Code formatting

+++ PerfectXL.EPPlus.UnitTests/DataValidation/ExcelTimeTests.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OfficeOpenXml.DataValidation;
 
@@ -12,7 +9,7 @@ namespace EPPlusTest.DataValidation
     {
         private ExcelTime _time;
         private readonly decimal SecondsPerHour = 3600;
-       // private readonly decimal HoursPerDay = 24;
+        // private readonly decimal HoursPerDay = 24;
         private readonly decimal SecondsPerDay = 3600 * 24;
 
         private decimal Round(decimal value)
@@ -79,9 +76,9 @@ namespace EPPlusTest.DataValidation
         {
             // Act
             _time.Hour = 1;
-            
+
             // Assert
-            Assert.AreEqual(Round(SecondsPerHour/SecondsPerDay), _time.ToExcelTime());
+            Assert.AreEqual(Round(SecondsPerHour / SecondsPerDay), _time.ToExcelTime());
         }
 
         [TestMethod]
@@ -94,7 +91,7 @@ namespace EPPlusTest.DataValidation
             _time.Minute = 20;
 
             // Assert
-            Assert.AreEqual(Round(expected/SecondsPerDay), _time.ToExcelTime());
+            Assert.AreEqual(Round(expected / SecondsPerDay), _time.ToExcelTime());
         }
 
         [TestMethod]
@@ -115,7 +112,7 @@ namespace EPPlusTest.DataValidation
         public void ExcelTimeTests_ConstructorWithValue_ShouldSetHour()
         {
             // Arrange
-            decimal value = 3660M/(decimal)SecondsPerDay;
+            decimal value = 3660M / SecondsPerDay;
 
             // Act
             var time = new ExcelTime(value);
@@ -128,7 +125,7 @@ namespace EPPlusTest.DataValidation
         public void ExcelTimeTests_ConstructorWithValue_ShouldSetMinute()
         {
             // Arrange
-            decimal value = 3660M / (decimal)SecondsPerDay;
+            decimal value = 3660M / SecondsPerDay;
 
             // Act
             var time = new ExcelTime(value);
@@ -141,7 +138,7 @@ namespace EPPlusTest.DataValidation
         public void ExcelTimeTests_ConstructorWithValue_ShouldSetSecond()
         {
             // Arrange
-            decimal value = 3662M / (decimal)SecondsPerDay;
+            decimal value = 3662M / SecondsPerDay;
 
             // Act
             var time = new ExcelTime(value);
