+++ PerfectXL.EPPlus/DataValidation/ExcelTime.cs
@@ -30,9 +30,6 @@
  * Jan Källman		                License changed GPL-->LGPL  2011-12-27
  *******************************************************************************/
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Globalization;
 
 namespace OfficeOpenXml.DataValidation
@@ -153,7 +150,7 @@ namespace OfficeOpenXml.DataValidation
         /// <summary>
         /// Hour between 0 and 23
         /// </summary>
-        public int Hour 
+        public int Hour
         {
             get
             {
@@ -234,7 +231,7 @@ namespace OfficeOpenXml.DataValidation
             var result = Hour * SecondsPerHour;
             result += Minute * SecondsPerMinute;
             result += Second ?? 0;
-            return (decimal)result;
+            return result;
         }
 
         /// <summary>
@@ -244,7 +241,7 @@ namespace OfficeOpenXml.DataValidation
         public decimal ToExcelTime()
         {
             var seconds = ToSeconds();
-            return Round(seconds / (decimal)SecondsPerDay);
+            return Round(seconds / SecondsPerDay);
         }
 
         /// <summary>
